examples nearly.
